requirements.txt
fastapi==0.92.0
keyboard==0.13.5
MarkupSafe==2.1.2
numpy==1.24.0
openpyxl==3.0.10
pandas==1.5.2
python-dotenv==1.0.0
pywin32==305
tqdm==4.64.1
